UTILS: make users use load_xstable_model
The support for XSPEC table models in the UI load_table_model command
has been marked as deprecated snice the 4.9 release but no-one knows
this (unless they read the documentation). So we now are explicit
with the deprecation warning (making it an actual logging call as well
as a deprecation warning).

The load_table_model call has been cleaned up (and it will make removing
the XSPEC support easier as we can now just delete the lines).

There are also a number of pylint changes, mainly

- explicitly catch Exception rather than a bare "except:" block

  lint-wise this does not save much, but it does mean that we no-longer
  catch the "exit exceptions" like Control-C, which should be better
  for responsiveness and makes the code clearer about what we are doing.

- use f-strings

- except for logging statements where we do lazy logging

- remove one unused symbol