IO: cleanup of the backend code
Mainly add typing information and make some internal changes to how
the code access the information from FITS files. This is not a complete
refresh, but does address some issues (although the interface is about
to be changed quite significantly anyway, which will remove some of
these changes).

One change is that the pyfits backend should now close any FITS
file it opens - previous to this commit some code paths would not
explictly close the file, leading to random "file not closed" warnings
on exit (e.g. from tests).

The crates backend has been changed to avoid the need to load in
the cxcdm module to set the TLMIN4 keyword for a RMF file, as it
turns out you can just set the keyword directly (it is never clear
how these special-cased keywords are handled).

There are some other changes too, such as

- pyfits: no-longer special case COMMENT/HISTORY/blank keyword handling
  since the situation the code is looking for (i.e. sent a sequence of
  strings rather than a string) doesn't actually happen in the code and
  so the existing code ends up writing out a HISTORY line as multiple
  one-character HISTORY lines

  (we don't have a good story about handling these keywords but this should
  be handled at the generic layer first, if we care about doing so)

- pyfits: there was an attempt to set the name of a HDU if not set
  when writing out tables but:
    a) why do this (it used the name "HISTOGRAM" which is not a
       generic name for a tabular dataset)
    b) recent changes I made actually made this somewhat pointless
       since it was changing the "tbl" object after we'd converted
       to the "hdu" object (ie the code wasn't actually helping)

- pyfits: allow for the case that the fits.HDUList object was created
  directly rather than read in from a file, and so has no "_file"
  setting (or, rather, it has one but it's None).

  I note that the current test suite does not cover this use case
  (that is, sending in a fits.HDUList, pycrates.TABLECrate,
   pycrates.PHACrateDataset, ... instead of a string), and it's unclear
  exactly what we support, so I expect there could still be problems.