IO: add documentation and typing for the dummy backend
Ensure that the dummy backend can be taken as a definition of the IO
backend "API" (as much as possible). The typing rules are a "best
effort" since there are still a lot of work to fully add typing to
Sherpa. To try to ensure that the types are shared, this creates
the sherpa.astro.io.types module - intended for internal use.

Note that this does not document the packup arguments as I plan
to remove them. This is partly because the code as currently
stands is "somewhat ugly" from a type perspective since we
have something like

  @overload
  def set_table_data(..., packup: Literal[False]) -> None:
     ...

  @overload
  def set_table_data(..., packup: Literal[True]) -> Any:
     ...

and it would just make sense to separate out the two functions.

There has been discussion about turning the IO code into a
module-based system, as is done for the plot backend, but that is a
significant internal change for only a small improvement in code
re-use (there was an attempt to do so during development of this code
and it was quite a large change). It could still be done, but having
documentation here would help that goal.

Fixes documentation typos noted by @hamogu during development of
this code.