IO: separate pack and set commands
This splits the old set_xxx_data code into two:

    pack_xxx_data

which only creates the "object", and set_xxx_data, which calls
the pack routine and then write the file out. This means we can
remove the packup argument, which was the cause for the confusing
return type, as well as no-need to send in irrelevant information
when all we care about is the pack call.

As part of this the backend code is now expected to provide
pack_xxx_data calls which are not called by the generic IO layer;
for example pack_arf_data.

This only really affects the I/O backend layer, as the generic
I/O layer hides this from the user, by providing access to the
pack routines and the set routines separately.

The typing of these routines is a "best effort" since the plan is
to change the more-difficult arguments to more-structured types,
so it doesn't seem worth the effort.