citation: handle Zenodo breakage
It looks like the recent Zenodo "upgrade" broke the existing API
for querying data. This commit fixes this, but it could happen
again and it is not clear we can do anytng about it, e.g. the Zenodo
API isn't versioned so we can't force a particular version and
even if we could, we have evidence that they can break it.