CI: drop ds9 tests on macOS (temporary)
Our tests can fail on macOS due to ds9-communication issues, but
it is not-at-all-clear why. It is also not repeatable (although
it happens many times), with the belief that it depends on what
system/setup the macOS runner uses.

So, for now just drop the ds9 tests by just not installing ds9
on the macOS system (we could probably also drop installing
xquartz but I have not gone that far). It means tweaking the
smoke tests (so it's not looking for DS9 on macOS).

Includes suggestions from @Marie-Terrell