IO: add HEASARC-specific table blocks
This change adds specialized TableBlock classes for the
PHA, ARF, and RMF file cases. These specialized classes are
meant to help avoid repeating checks we know should pass (e.g.
that a PHA dataset has a CHANNEL column) but it doesn't actually
add much type safety. It does add some level of documentation
to the backend code (i.e. point out what structures we expect,
rather than just a generic "table").

One issue is that these clases raise a generic ValueError
error rather than a DataErr (since, in part, we don't know
the filename when throwing the error, because they could be
generated from in-memory data). I am assuming that this is okay
(since, for most cases) such errors should already have been
caught in the backend code.

An alternative would be to add special-case classes with
direct access to the columns we require (e.g. as a dataclass),
but this way we allow for extra information (e.g. other columns)
to be sent around.