Support multi-matrix RMF
Provide initial support for multi-matrix RMF. There are basically
two approaches given the current structure

- create multiple DataRMF files and apply them individually
- create a single RMF instance that applies them all at "once"

The former would perhaps fit into the existing code, since we have
the existing sherpa.astro.instrument.MultipleResponse1D class, but

a) I don't really understand the trade offs that has
b) the existing code deals with multiple files, where it makes
   sense to incorporate them, whereas the multi-matrix case
   has a single file

So I have gone for the second option as an experiment. This also
*technically* has the advantage that we can make this more efficient
as we just call the rmf fold once, and there can take advantage of
knowledge of the response energy grid. However, as discussed in
issue #1907, we don't have a good way of associating the evaluated
model with the energy grid used to create it. So we need to work
on this (which will probably lead to better understanding the
MultiResponse1D class as well).

There is also the fact that the DataRMF class has special case
for filtering the data based on the noticed channels in a PHA
which we currently overwrite/avoid in the multi-matrix case. This
is again because I do not understand this code.