models: table models check for numeric axes
We now error out if we can not add 0 to either the independent axis
or, if givenm, the dependent axis sent to a table model. This check is
not intended to catch all problem cases, just the "common" case that
caused #1943, where a column was all strings.

To do a more-extensive check would require either testing other
operators (e.g. multiplication, division, subtraction, power, ...)  or
comparing against the number.Number "class". However, there is always
the case that a user could have created some particular numeric
variant that passes any test we write, and so - as this is not the
most-common error we see - it does not seem worth trying to be "too
clever".

I explicitly chose not to check that all elements of the array, or
arrays, pass "numpy.isfinite" as it is possible to want to read in a
file containing NaN/Inf values for a table model and then
automatically exclude these from analysis (thanks to a notice or
ignore call).