Tests: compare Response1D and MultipleResponse1D behavior
Add basic tests like

- Response1D and MultipleResponse1D are the same for a
  single-response PHA
- adding the same response twice for MultipleResponse1D does
  double the signal sen by Response1D

In doing so I have re-worked some existing Response1D routines
so they all call setup_resp() rather than repeating this code
in each test.

I will note that the MultiResponse1D tests are currently
meant as regression tests since I have not had time to really
understand what the code is meant to be doing, and understand why
it appears to behave differently to the Response1D case when
we have the expoure time coming from the ARF and not the PHA
(which is post-OGIP behaviour, as noted in issue #1871). That is,
I would not be surprised if we decide to change some of the
behaviour. I will note that the string name for the multi-response
case is definitely wrong here.