BUILD: initial support for meson
With this I can

    pip install .

and then in another directory

    sherpa_test -n auto

Editable installs must use

    pip install -e . --no-build-isolation

which requires that meson, meson-python, ninja, and numpy are
installed in the user's environment.

There is some support for building optional code, but

- FFTW currently requires that it is provided by an external
  entity as updating the local build to meson is currently
  above my pay grade

The extern/ code has been moved to subprojects/ since it feels
like using the meson subproject support makes sense here.
However, this is just a vibe and I may change my mind. Several
thoughts:

- I haven't removed/cleaned up the existing configure support
- although they are labelled subproject, the build is very
  tailored to Sherpa's needs (which may suggest going back to
  just make them part of the general build)
- however, this way we can get per-subproject C standard
  support, such as wcssubs needing -std=g<v> rather than -std=c<v>
- labelling of the subproject directory: do we want to keep the
  version number or not

The editable build now requires the --no-build-isolation flag,
which complicates the build somewhat (as we need to ensure we have
the build components). There are definite improvements when
built like this (I was able to change the code for an extension
module and then the pytest call picked up this change and compiled
just the needed code).

The existing setuptools code has also not been removed.

Initial documentation support has been added, but more is likely
needed.

Version support is incomplete (e.g. the replacement for
sherpa/_version.py and its interactions with the docs build).
There is an assumption that the python3 executable is available,
but this does seem to cause problems with some editable build
changes.