meson: first partial support for meson
With this I can

    pip install .

and then in another directory

    sherpa_test -n auto

and the tests pass *EXCEPT* that many optional modules are not built,
so the test coverage is low.

It also is hard-coded

a) to a particular version

   How do we handle the version using meson-python?

b) the location of the fftw include and library files / locations

   How do we pass through information? Presumably we can just
   use the existing setup.cfg file since this already exists, even
   if it's not a meson-python native solution.

The existing setuptools code has also not been removed.