Tests: better support of optional region/wcs dependencies
The idea is to allow Sherpa to be built withut the region or
WCS libraries. These are low-level Astronomy-specific modules so
it makes sense to make them optional, but we don't currently
make this easy to do so (that is, this commit does not make it
easier to avoid building these two extension modules).

This is mainly changes to the tests, to add decorators, but there
are code changes

- addition or requires_region and requires_wcs decorators
- the sherpa.astro.io.wcs module has been changed to make the
  lack of the WCS code an option, rather than being required
  There are other ways to do this, but for now it's easiest to
  do it this way, as it depends on the long-term build changes
  that are planned.

This was work I did to help with cnoverting to Pyton 3.12 (the
distutils changes) but it turned out not to be necessary for that.