build: move to recent setuptools
Move the configuration to pyproject.toml from setup.cfg and
deal with changes in distutils support (Python / NumPy).

Bump minimum setuptools to 64.0.0 since

- this allows for editable installs following PEP 660
- there are a lot of changes in setuptools and it's hard to test them,
  so using a slightly-newer minimum version avoids testing / debugging
  of old code; 64.0.0 was released Aug 11 2022

The MANIFEST.in file could probably be cleaned up, but I do not
have the energy and it can also depend on the version of setuptools
in use.

I still do not understand how to mark files for inclusion - whether
as packages or data files - and the current code seems to work for me,
even though it provides warning about sherpa.xxx.src directories (which
I'd like to exclude but do not know how) and is also likely excessive
in the tools.setuptools.packages setting. This might be simpler if
we moved to a src/ layout.