TEST: add checks of IO "object" handling
The code supports, but does not test or well advertise, passing
around the "objects" you get from the chosen I/O backend rather
than a filename. This commit adds some basic checks of this
support.

There is some "interesting" difference in behaviour for PHA files
between the two backends (I did pick a particularly-interesting
test case for this, as it's a CSC source).