IO: clarify routines and clean up
This is partly addressing review comments by Moritz.

Adds documentation to the dummy backend to describe the arguments.

Fixes up differences in the typing of routines in the dummy,
crates, and pyfits backends.

Adds typing to some of the routines in __init__.py.

A few minor code changes (mainly style) to the code. Some of the
changes are to avoid issues with mypy: mainly creating new
variables rather than over-writing existing variables to avoid
confusing the type checker, but there is also some assertions that
have been added to ensure a backend has been selected (thanks to
the way we document the symbol) and at least one explicit error-check
if None case to avoid mypy unhappiness.