IO: make the backend class, rather than module, based
A new routine has been added to the sherpa.astro.io module,
set_backend, to make it easier to switch backend. This copies the
approach taken with the plotting backend.

The implementation is placed in backends.py as the BaseBackend
class, and the backends are expected to derive from this. That
is, the backend is now a class rather than a module (even though
the current design only has one class per module).

The actual implementation of the backend code - in particular
selecting the backend - is a lot simpler than the plot case because
the way the backend is selected here is that it requires the
class to be called Backend in a file called xxx_backend.py (and
have the name attribute of the class set to the string matching
xxx, e.g. "crates"). We could use the plot approach but it does
not seem worth it at present.

As part of this I've tweaked the ArgumentTypeErr "backend" error
added for the plot case (the label is now "tempbackend" rather
than "tempplotbackend"), as this is unlikely to have been used
in downstream code.

A lot of the test changes are because these tests need to know
what backend is in use, and the way we can check this has
changed. Previously I tended to use the symbol
sherpa.astro.io.backend.__file__, since sherpa.astro.io.backend
was a module. It is now a class and has a name field, so we can
use sherpa.astro.io.backend.name instead.

There has currently been no attempt to add useful common logic
to the backend class. This will come in a follow-up commit.