IO: move common logic into the base backend
As suggested by Moritz, move common logic into the base backend.
It is a bit messy, since the base backend is mostly abstract, so
we end up having to create a new method (write_arrays) which the
backends are supposed to implement, with the set_arrays method
basically being defined in the base backend now.

There are some small cleanups in the tests because of this, because
the behaviou - in particular of the pyfits backend - has now
changed (we add a comment line rather than not writing one).