Tests: add tests of bug #1561
The interval-projection and interval-uncertainty commands do not
actually log the values when told to.

Note that we have two issues:

a) the x axis is not logged
b) the min/max values are logged for some reason

We have tests of the class and also tests of the UI layer (which
technically should not add any extra power to the tests but is
closer to the actual bug report in #1561).

There is also some minor clean up of the existing tests and coverage
of some corner cases (mainly error handling).