Confidence: improve handling when delv is set
The delv parameter lets the user specify the spacing, rather than the
number, of the parameter values. The old code appeared to have a
mis-understanding of how numpy.arange works for the upper limit, and
can create values that are greater than the requested maximum value.
This is wrong and easily fixed.