image: add documentation to the backends and minor clean up
Add consistent documentation to the dummy and ds9 backends. This
includes changing the dummy backend from accepting (*args, **kwargs)
to the actual API for each call. The pyds9 backend is not changed,
other than adding a very-basic introduction, because

a) pyds9 is no-longer maintained
b) we do not test this interface

The _set_wcs command has been moved out of the dummy backend as it
really is just something for the actual backend to decide to implement
or not. Note that there is shared code between the ds9 and pyds9
backends - in particular related to _set_wcs - but I have not tried
to move this into a commin location as I expect the pyds9 backend to
be removed (or at least sun-setted).

There are some small changes, suggested in part by pylint.