plots: add a MultiPlot class
Support plotting multiple datasets as a single "plot". This is
currently only used for the source_components and model_components
style plots but could be used for

- order
- fit
  - although I am unlikely to suggest it for this case as the
    existing fit support (e.g. FitPlot) is likely well used by
    the community and not worth changing

This also adds support for

  plot("model_components", "source_components")

This adds a "set_title" feature to the plot backend code. I assume
that the bokeh changes are correct, and that this approach is
sensible. We add explicit matplotlib and bokeh tests that checks
this set_title feature.

The MultiPlot class handles either plot or histogram plots: that
is anything that

a) uses the prepare method to create the data to be used in
   the visiualization, and that this step has been performed before
   creating the MultiPlot object (although technically this is
   not necessary; it is only necessary to have been done before
   the plot method is called)

b) uses the plot method to create the visualization

We could add a MultiContour class to mimic MultiPlot, but we do
not have any use case for this at the moment, so I have decided
not to add one.