plot: add typing rules to the backend code
Adds typing rules to the backend code:

a)

Change the implicit typing of the backend variable from
"Optional[BaseBackend]" to "BaseBackend" by moving the creation
of the variable to after the basicbackend variable is set up. This
means down-stream code doesn't have to worry about backend maybe
being a None value.

b)

Add typing rules to the set_backend code. There's also some small
style changes (e.g. try/except rather than "if key exists ...").

c)

As part of b, the context-manager typing changes pointed out that
our context manager wasn't quite correct - e.g. we did not return
ourselves and we were not explicit about the exit statement.
So fix this. It does not actually affect our code, but let's be
correct in case we change the cnotext manager at some point in the
future.

d) [not backend related, but is a typing change]

We have float + numpy-float which mypy dislikes. The easiest way to
address this is to explicitly convert the numpy-float to a float.
This is fallout from #1959 (i.e. I thought I'd done it there but
obviously not).