plot: add BasePlot, BaseHistogram, and BaseContour class
We have the base classes Plot, Histogram, and Contour which are
displayed via the plot, histogram, and contour methods, where
the method is sent the actual data to display (e.g. x and y for
the plot cases, xlo, xhi, and y for the histogram case, and
x0, x1, y - along with the levels - for the contour case).

We then derive classes from these where the aim is that the user
calls the prepare method - which extracts the data to plot from
the various Data/Model/Stats classes - and then the plot,
histogram, or contour call does not take the actual data values
(as these have been stored internally by the prepare method).

For the histogram class we have the HistogramPlot class that sets
this expectation, and all the remaining histogram classes derive
from it. However, we do not have the same structure for the Data
and Contour classes. So here we introduce the BasePlot and
BaseContour classes which do this. It would be nice to match the
histogram case and use the name DataPlot rather than BasePlot,
but this class already exists. So we rename the HistogramPlot
to BaseHistogram - to match the other new classes - and then
have HistogramPlot just be a re-export of BaseHistogram.

Unfortunately we can not easily use the abstractmethod property
to label these classes for the routines that need to be provided
(prepare and _repr_html) because I do not want to go and add
ABCMeta into the class hierarchy (given some of the changes I
note below with the use of multi-dispatch).

The __str__ method for these classes is essentially the same
for these plots - we have multiple lines of 'label = value' -
and so this logic is moved out into a helper routine
(display_fields) and the choice of fields to display is left to
the class (via the class _fields setting). This could have
been done via an abstract base class, but that would require
invasive changes (inheriting from abc.ABC) so we currently have
repeated code (as it's not excessively large). One change because
of this is that the string output is more uniform (for lines like
"label = value" the spacing is a lot-more consistent but it means
there are some small changes in the number of spaces).

Note that the __str__ behaviour is similar to how it is handled
in the sherpa.data.Data hierarchy, but we do not try to share
code because it does not seem worth it, as the outputs are subtly
different. We perhaps should standardize, but that would be a
different PR.

The recent changes to make more use of super() has led to small
changes in the choice of labels (e.g. is "xlabel" set to "" or
"x"). In general I think the changes are better (and simplify the
test code slightly as the behaviour is more obvious). It is not
all sunshine as there are a number of cases - particularly in the
sherpa.astro.plot module - that have multiple super-classes,
and this has lead to some issues over the string representation
(which has meant that some of the astro-specific commands require
explicit _string_fields settings to over-ride the MRO default).

A careful review of some classes - such as the residual-style
classes, but some of the data classes too - show that the decision
whether to have/display the xerr and yerr fields is somewhat
random. This results in `plot` methods not being able to call
the super-class but having to hard-code the parent class.

I do wonder whether the BasePlot/Histogram/Contour classes should
just inherit from object, and instead contain a
Plot/Histogram/Contour object (i.e. a has-a rather than is-a
relationship). This would be quite a conceptual change but would
clean up some of these issues (but I am not clear whether it would
generate other issues).