plot: add BasePlot, BaseHistogram, and BaseContour class
We have the base classes Plot, Histogram, and Contour which are
displayed via the plot, histogram, and contour methods, where
the method is sent the actual data to display (e.g. x and y for
the plot cases, xlo, xhi, and y for the histogram case, and
x0, x1, y - along with the levels - for the contour case).

We then derive classes from these where the aim is that the user
calls the prepare method - which extracts the data to plot from
the various Data/Model/Stats classes - and then the plot,
histogram, or contour call does not take the actual data values
(as these have been stored internally by the prepare method).

For the histogram class we have the HistogramPlot class that sets
this expectation, and all the remaining histogram classes derive
from it. However, we do not have the same structure for the Data
and Contour classes. So here we introduce the BasePlot and
BaseContour classes which do this. It would be nice to match the
histogram case and use the name DataPlot rather than BasePlot,
but this class already exists. So we rename the HistogramPlot
to BaseHistogram - to match the other new classes - and then
have HistogramPlot just be an alias to BaseHistogram.

Unfortunately we can not easily use the abstractmethod property
to label these classes for the routines that need to be provided
(prepare and _repr_html) because I do not want to go and add
ABCMeta into the class hierarchy (given some of the changes I
note below with the use of multi-dispatch).

One interesting change is that now FitPlot derives from BasePlot
rather than Plot, which means it gains a number of attributes
it does not need/use, and that can confuse downstream code. In
this case JointPlot needed to change it's plottop behavior
to check for plot.data/modelplot.xlabel before plot.xlabel
rather than after it. This suggests that the BaseXXX class should
perhaps just define the prepare/{plot | contour} methods and
there is a different class for the attributes.

A careful review of some classes - such as the residual-style
classes, but some of the data classes too - show that the decision
whether to have/display the xerr and yerr fields is somewhat
random. This results in `plot` methods not being able to call
the super-class but having to hard-code the parent class.

I do wonder whether the BasePlot/Histogram/Contour classes should
just inherit from object, and instead contain a
Plot/Histogram/Contour object (i.e. a has-a rather than is-a
relationship). This would be quite a conceptual change but would
clean up some of these issues (but I am not clear whether it would
generate other issues).