model: simplify model expression names
Remove excess brackets from model names. The idea is to reduce the
number of brackets in typical model expressions, not to

- remove all un-nescessary brackets
- simplify model expressions

So, the aim is not to necessarily simplify all expressions, just
to handle the most-common cases.

If you look "in the literature" (a.k.a. a quick google search) you
find a lot of discussion about converting to RPN / postfix form. We
are lucky because the model expression structure is easy to process
in a RPN-like form. In fact, all we really need to worry about is
the BinaryOpModel form (we could also include UnaryOpModel terms
but for our use case it doesn't seem worth it). So we can use the
normal approach of

    process the lhs
    process the rhs

and then decide whether to add brackets around lhs or rhs depending
on the current operator and the operators in the lhs or rhs terms
(if any). Now, subtraction and exponentiation complicate things a
bit, so the logic is less clear than you would hope.

This is why there is an extensive test case which checks the name
field returns the expected value. This set of tests attempts to

a) cover common use cases
b) cover the code paths

The code was not written to be short or particularly performant.

One potential concern is that the algorithm is recursive, so it
is possible for the code to error out with a RecursionError. I
would like to be able to catch this case, since we could just use
a less-simplified result, but it's not obvious that we can do so.