Fix the requires_pylab decorator
As noted in #1971 the current requires_pylab decorator, which is
meant to only run tests when matplotlib is available, and then
run the test with the pylab plot backend selected, does not work.

This is because it's trying to be both a decorator and a generator.

The easiest way I have found to do this is to move from requires_pylab
being a decorator defined in sherpa.utils.testing to being a pytest
fixture defined in sherpa/conftest.py

I think there's a strong argument to be had to make these tests
fixtures rather than decorators, but I am only doing it for
requires_pylab because

- it is the only one that needs this
- there are not many users of requires_pylab so it is an easy-to-make
  change
- requires_pylab is a fairly-new invention so there's less likely to
  be downstream problems with this change
  (since it is simpler to remove the sherpa.utils.testing.requires_pylab
   symbol to make it obvious what is going on)
- doing the other symbols
  - would take too long
  - although I think it would be better, there may be some cases where
    we would want to retain the decorator

As part of the test changes - since they haven't been run for a while,
and so hace fallen out of date - one contour test triggered a failure
when run with matplotlib 3.8. The way contour data is stored has changed
in matplotlib 3.8. As it's unclear what's the best way to get this data,
and what we are really checking, let's just make sure there's something
that's reported.