Tests: add tests that show off issue 777
The issue is that we currently require all fittable parameters to be
explicitly included in the model expression, which means writing
model expressions like

    mdl1 + 0 * mdl2

when all we want is to use a component from mdl2 in a linked parameter
of mdl1.

So add some tests that check this. We add low-level ones and UI level
ones (although for the UI case we do not add the
do-not-add-a-linked-parameter example, as we assume we have plenty of
tests that check this for the UI case already!).

We can see that when we do not include the model explicitly the
fits complete, but not to anything a user cares about. However, since
it's technically fitting thngs, we label these as "passes" (so we can
see if we ever address issue #777).