Tests: add checks of stringification of parameter expressions
This is similar to the model case, but the examples are slightly
different (less worry about cnovolution-style cases like RMF here).

There's also a test of a string expansion which I do not feel is
correct - the expression

    (-p) ** 2    # where p.name = "m.x"

gets mapped to the string "(-m.x ** 2)" but, thanks to precedence,
this is actually -(p ** 2). Rather than mark this as XFAIL I
explicitly test the current behavior.