parameter: simplify parameter expression names
Follow the same approach as used with model names. So this is an
attempt to remove excess brackets without

a) actually simplifying the expression (so "(a - a)" becomes "a - a"
   and not "0"

b) it does not simplify everyting (so "-a" actually becomes "-(a)")

Note that (as highlighted) there is a small reduction in expressivity
of the langage when applying a unary parameter to a single term, we
now always add the bracket (so instead of -mdl.c0 you get -(mdl.c0).
I believe this is a price we can afford to pay.

Helper routines have moved from sherpa.models.model to sherpa.utils
and the typing changed (use of Any rather than something like
Union[Model, Parameter] to avoid forward links).

Now, because of the changes in #1653 the stringification of
parameters is actually more-complex than of models, but fortunately
it doesn't seem to be a problem for the approach I developed for
the model case (which is good as I want to look at implementing
this change for models).