model: move operator precision into the *Op* classes
Rather than always querying for the precision, we store it along
with the operator and (maybe) string. This lets us simplify things
a little bit (we can ask for the precedence directly if the object
is one of the types that we care about). I have also moved the logic
for determining whether the left or right sides of the expression
should be surrounded by a bracket into the sherpa.utils module. This
moves the logic away from the place where it makes sense, but there's
no point in having the same code repeated twice if we can help it.