Tests: explicit fit test of duplicating a model term
If we use the same model multiple times then do we get the correct
answer? That is, we do not treat the fit as having extra parameters
and the parameter results end up correct?

There is likely a test of this somewhere but it's hard to find
so add one (it re-uses the new linked-parameter tests).

We also check on the meaning of the pars field for a "complicated"
model to check

- linked paramters
- what happens to repeated model parameters

as it's not really obvious from the code, so we want a check of the
existing behaviour.