model: get_thawed_pars/thawedpars now includes linked pars
This is to fix #777

The thawedparXXX attributes rely on get_thawed_pars() to select
the parameters. Prior to this commit get_thawed_pars only
returned the parameters that are incuded in the model expression,
that is essentially we had

    mdl.get_thawed_pars() == list(mdl.pars)

With this commit, we mnow have

    mdl.get_thawed_pars() == list(mdl.pars) + list(mdl.lpars)

This means that we now have access to linked parameters, and so
fit will now include them even if they are not included in the
model expression. This is thanks to the change to using get_thawed_pars
throughout the code base, rather than explicitly looping through the
pars attribute.

We do not want to change the elements of the pars attribute to include
these linked parameters as this would really complicate the model evaluation
code, since with the current design we know that self.pars gives all
the parameters for the model evalaution, and BinaryOpModel can split
this between the LHS and RHS terms. I did try carrying around the number of
parameters each model actually needs, so we can split out the linked terms,
but it was too costly (as models need to worry about this when writing
their calc method which is just confusing for the model writer and would
cause problems when users try to use models written before this change
was made).

Fortunately the only other change needed is to some utility routines in
sherpa.ui.utils (and most of the changes are actually typing related),
since the sherpa.fit code was changed to use the get_thawed_pars routine
earlier to automatically pick up this improvement.

I note issue #1981 here - which was found when developing this code. As
this problem existed before this PR was made I have not tried to fix it
(since it is not-at-all obvious what to to).