ui: internal improvement for joint plots
Alhough not an issue just yet, the joint-plot code assumes the
second plot area is a "data" rather than "histogram" plot.
Specifically, it assumes the plot object has a plot_prefs field
when the plot objects for the top plot allow either plot_prefs
or histo_prefs.

The solution is easy - we can use the same routine as used for
the top plots.

For the astro version it is a bit different since we know that
the plots in the fit/model case are integrated cases (DataPHA)
and so we could avoid this logic. However, the residual case
may change, so make it flexible and look like the sherpa.ui
version.