plot: histogram-style residual plots
Data and model plots support specializations for Data1DInt and
DataPHA classes, so extend this for the Delchi, Resid, Ratio,
and Chisqr plot cases.

The naming of the plot classes and the duplication in code is not
ideal, but I leave that for later to address (or just identify as
an issue).

This means that users lose the xerr field on residual-style plots
for PHA analysis. Is this something we care about. This field was
used in the original report for issue #1817, so users may

a) find their code breaks
b) may miss this feature

if they are using the plot structures for their own ends. An
advantage to it breaking code is that users will have to look at
the structures and (hopefully) find the xlo/xhi attributes are
now available.