plot: internal re-work of PHA plots
Move the logic for calculating the low and high edges of each
bin (X axis) out of the classes and into a separate routine.

Technically this logic could be moved into a parent class or a
helperclass, but the current class structure is "rather messy" and
I do not want to address that here.